+++ diskover_qumulo.py
@@ -149,6 +149,7 @@ def qumulo_treewalk(path, ip, ses, num_sep, level, totaljobs, batchsize, cliargs
             if len(batch) >= batchsize:
                 diskover.q.enqueue(diskover_worker_bot.scrape_tree_meta,
                           args=(batch, cliargs, reindex_dict,))
+                totaljobs.value += 1
                 del batch[:]
                 batchsize_prev = batchsize
                 if cliargs['adaptivebatch']:
@@ -177,6 +178,7 @@ def qumulo_treewalk(path, ip, ses, num_sep, level, totaljobs, batchsize, cliargs
     # add any remaining in batch to queue
     diskover.q.enqueue(diskover_worker_bot.scrape_tree_meta,
               args=(batch, cliargs, reindex_dict,))
+    totaljobs.value += 1
 
 
 def qumulo_get_dir_meta(path, cliargs, reindex_dict, redis_conn):